requirements.txt
streamlit
pymongo
pandas
vaex
python-dotenv
scikit-learn